chore: delete unused file and imports
